bloom.go

+++ bloom.go
@@ -32,7 +32,7 @@ func (b *BloomFilter) Add(data []byte, p float64, q float64) *BloomFilter {
 	adr := uint32(2147483647)
 	for i := uint32(0); i < uint32(b.numHashes); i++ {
 		trueBit := ((uint32(lower)+uint32(upper)*i)%uint32(b.filterSize))
-		for j:= uint32(0); j<b.filterSize; j++ {
+		for j:= uint32(0); j<uint32(b.filterSize); j++ {
 			var r uint32;
 			if q==1 && p==0 {
 				// 0.5 is an arbritrary number which is less than 1 and greater than 0
